Optional: Monitor logfiles with Loki.
To monitor also logfiles the Loki stack is used.
That means we use Loki which acts like prometheus but is for logs. This time the monitored logs need to pushed there, which we are doing with promtail. This is a lightweight client, which can tail log files and sends them to loki.
In Grafana we can explore the collected logs after adding a new datasource und also pin them to our dashboards.